Add in config handling subgroup
Adds in a `config` group that provides commands to get value(s), set a
value, and remove values.  This will also handle creating the required
config file if it doesn't exist and works for both the local repository
and the user level configs.  Some rearchitecting occurred for how a
module gets bound to the parent group (this will probably be changed
later to a more decorator specific format) and how config files were
denoted and set so that the config system could target modifications for
specific files.