Add in alias support
Alias support is a bit half baked but works.  It adds a new section to
the config that allows for "aliases" to be defined, which end up
effectively being string expansion for the alias into what it is defined
as.  This means that setting an alias of `crst` as `create --start`
would mean that running `den crst` would expand to `den create --start`.
Any additional arguments will be carried over into the expanded command.

This creates a new Group class that performs the expansion if normal
command resolution failed (this means you can't alias standard
commands).  It also leverages the calls for setting/getting values in
the config command set for the `alias` command.  There are some issues
with this as it means there is not "global" get that uses all config
files, but this will probably end up being another pass.