setup.py support
Work involved with getting setup.py working.  Added some make stuff to
work with this and it's artifacts, added a README for the command sets
folder (just to help with documenting the process), added a decorator to
give boilerplate docker error support (so catches docker errors at raised
inside the decorated function and provides a clean error).