Linting+Testing introduction
Implemented an initial pass on testing and linting.  Cleaned up the
codebase to pass the linter (currently skipping the tests).  Added a
basic pass on testing and a framework for easily adding additional
tests.  Cleaned up the requirements files to better dictate the base
packages for both regular usage and for testing.